First move of serving reconcile sytle.
